Fix: Setting memory cache size greater than 2GB causes a segfault.
The problem was in the block_address() function that returns the memory address
for a given cache block, It was using 32bit integers to calculate the offset of
the block within the shared memory space that is only good until the 2GB limit.

+++ src/query_cache/pool_memqcache.c
@@ -3025,7 +3025,7 @@ block_address(int blockid)
 	char	   *p;
 
 	p = pool_memory_cache_address() +
-		blockid * pool_config->memqcache_cache_block_size;
+		(uint64)blockid * pool_config->memqcache_cache_block_size;
 	return p;
 }
